Remove the network constraint from the WorkRequest.
While the request does technically need a network connection to succeed,
asking Android to enforce it needlessly results in about a 3 second
delay between each task. If the task is run without a network
connection (or with an intermittent one), it will quickly fail anyway and ask
to be retried.

Signed-off-by: Jason Lynch <jason@calindora.com>