Move signature generation and URL determination into the actual worker.
Previously, by generating the signature at the time of the report, the secret
key was finalized into that signature. If the user had made an error in the
secret key (or the URL or the device key, which were also finalized), that
incorrect data would be permanently with the task, leading to it being unable to
ever complete, and therefore clogging the work queue.

By deferring reading the URL and generating the signature until just before the
attempt, the worker can always use the most recent data, allowing the user to
clear the queue by fixing any configuration mistakes.

Signed-off-by: Jason Lynch <jason@calindora.com>