Add a couple of debug options to help prevent submissions from getting stuck.
The first option is to clear the entire queue. This is gated behind a
second button reduce the likelihood of an accidental press. The second
option is to enable an option that will clear the first item in the
queue whether it succeeds or fails to submit. The use case for either of
these is if work is stuck.

There are a few possibilities why work might get stuck:

1) The user makes a mistake entering their device key or secret. In this
   case, the server will fail to commit. The user can fix this problem
   already by correcting their key or secret and future submissions will
   work correctly.

This first one is not the focus with this commit, but the following two
are:

2) There is a bug in the server that causes it to reject a valid
   submission.

3) There is a bug in the client that causes it to generate an invalid
   submission.

In either of these cases (if there are any left), there is nothing the
user can do, and being able to clear the submission queue of the errant
submission(s) could be useful.

However, in an actually released application, I would want something far
more robust. Since this application is merely for personal use, I'm not
too concerned. The recent corner case I fixed was the first error in the
client I've seen in several years.

Signed-off-by: Jason Lynch <jason@calindora.com>